Copy digitalWriteFast from OpenCR
I made a copy of the digitalWriteFast stuff that we added to the OpenCR board.
WIth this I also made a copy of the example program that was added to OpenCR.
I changed the name as f_... was already taken, so it is g_...
Plus I had it grab the current micros() around the different groupings of calls so we can see the differences in timing.

Looks like this version is not quite as fast as doing the registers, probably due to calls and/or maybe arrays?

Example Output
digitalWrite:5 Regs:2 digitalWriteFast:1
digitalWrite:6 Regs:1 digitalWriteFast:2
digitalWrite:6 Regs:1 digitalWriteFast:2
digitalWrite:6 Regs:1 digitalWriteFast:2
digitalWrite:6 Regs:1 digitalWriteFast:2
digitalWrite:6 Regs:1 digitalWriteFast:2
digitalWrite:6 Regs:1 digitalWriteFast:2
digitalWrite:5 Regs:2 digitalWriteFast:1