Serial.flush() - Speed up
Current version that worked, just simply waited for up to 5ms for the timer to write the data out.

In this version, I changed the timer to 3ms which obviously helps some, but in addition added a swTimer function which allows you to set the counter.  So in the flush function, I force the count to 1, which will imply that the write function will be called on the next millisecond timer.

I updated the whole path of code for flush, where instead of asking if the writes were done, I created a flush_tx function at each level and all of this work was done in the usbd_cdc_interface function, which file already knows (it created) the timer.

Without the timer update change I have a test case, where I basically ask an AX servo for it's current position, and print it out and then call Serial.flush.  I put calls in to micros() around the calls to Serial.write/Serial.flush and sum the delta times and with only the change from 5 to 3ms, the average time for the flush to take was about 2.8ms.  With the force to next ms, the average time was about .7ms...