Add support for SerailX.write that does not block
Setup to use semi RingBuffer that the write directly to the output
register if the TXE bit is set and the FIFO queue is empty.  This helps
reduce the number of times that interrupts happen especially with high
speed transfers.

The code is also setup to allow the user to set a pin as the write
enable pin, which when set will automatically set the pin on each write
and will automatically clear it when the transmit complete (TC)
interrupt happens.  So again Transmit of packets does not have to wait
for writes to complete.

Updated the DXL code (at least the DynamixelSDK) to set this option.

I have done testing to see how well this works.  This includes running a
few of the Dynamixel SDK example programs as well as a Dynamixel
Workbench app.

I have also run my own sudo version of default firmware on OpenCM9.04
board and was able to communicate with R+ Manager version 2, and tested
it talking to a few servos, mostly at 1mbs, but also tried 2mbs and
checked 3Mbs using logic analyzer.

Added keywords.txt - to allow some newer things like INPUT_PULLDOWN to show up as a know keyword.

updated files in opencm9.04_release including header files and keywords.txt, and lib_f103.a