Dynamixel Workbench: All register names exported on other platforms exported on OpenCM/CR
Copied over the proposed changes I did for OpenCR ,
Only changes to that is needed to edit the .cpp files to change the include file locations as the directory format is different between these two projects ?

Starting a little testing.  Right now appears to be a bit flaky, but I have had issues in past using the Monitor program as there does not appear to be very much error dection nor recovery in it.