Workbench update: update release headers and lib_f103.a
As I mentioned earlier I had not yet updated the headers and lib, as for example was not sure if the PR for the non blocking writes would go in first or not.  So the lib file would have needed to change again...

Also this copied the library .cpp files as well to release as they were there

Also updated that one sketch file that I updated.