Dynamixel Workbench - Reduce size of main object
When I started this edit, when I did a call like:
  Serial.print("Size of DynamixelWorkbench: "); Serial.println(sizeof(DynamixelWorkbench), DEC);

It was printing 2312.

With this edit, it now reports 512 bytes.

The main change is I removed the model_name and Model number out of the DXLInfo item which was for each item and was a character array of 20 bytes.

Instead I have the DynamixelTool hold onto a const char * pointer, which is returned from mapping IDS to strings.  The mapping functions were changed to a simple mapping table that has model number and a const char *...

I also tried to put a little more testing of values in place so if you pass in an ID that does not exist it tries to recover.

Also the code tries to now handle having more than 16 servos of the same type.  If it detects that a tool is full, it tries to find or create another tool to continue with that type of servo.

Also has a version of the Find Dynamixel that works more Arduino like... That is it prints using normal Serial functions and does not use String objects.  Also the loop code is in loop() and allows you to rerun the scans.