Build Develop branch on ARM Linux (RPI)
There were issues with building the develop branch on an RPI3.

The compiler would error out with utoa not being defined.

I found that on the Teensy 3.,x builds, that they conditionally defined this function, depending on if newlib was define _newlib_version < 2.2...

So defined it that way and the compile completed at least for blink on an RPI3.

Still issues with needing loader program for ARM linux...