EEPROM: Fixes - Only 255 items, don't write if same value, check max address
There are only 255 values that can be stored in EEPROM not 512.  The page alternate between each other for which one is valid.  And first bytes used for page status.

Also restrict the address passed in toe EEPROM.write(...) to less than the size, so only the size number of addresses can be stored, else could run into issue where there is not room and...

Also put in code to see if the value writing is the same value as was the last previous item written.  If so bail out and not write new value...

Updated lib_f103.a file with this as well.