EEPROM: Fixes and enhancements
This combined all of my EEPROM fixes from the other branch(s), which include:

a) Fix the PAGE1 issue which would cause us to write some stuff in bad page which caused us to have to recvoer the firmware to work.

b) More or less rid us of the VirtAddVarTab table, which was eating about 1K of user memory.

c) Limited number of variables we could store to 255. This was because the Pages that we write to are 1K and with 4 bytes per element we only had 256 minus the first one which had a page status value

d) Put check in to not allow us to write to random address (>= NB_OF_VAR).

e) Now since number of entries is < 256 and the EEPROM library only stores byte values I changed the code to save in two bye elements, which reduced how many times we had to copy pages back and forth.

f) If we do an EEPROM.write of some value and the EEPROM is already set to that value, don't write new entry for it.

Update lib_f103.a to these updates.

Note: I have another set of branches with this, but merged it all here to hopefully avoid the issue of needing to handle conflict of the lib_f103.a file