EEPROM Changes to 2 byte after 4 byte version merged.
Here are the later changes made to convert the 4 byte entries into 2 byte entries.

This change also included an optional change I made, which deferred the actual initialization of  the EEPROM until someone actually needed it.  That is until some one calls something like EEPROM.read or EEPROM.write, or  the size function.

Two reasons:   Why take the time to init if if the actual program never uses it.
And more for my point, if you init the cod eas early as drv_eeprom_init is called, you can not use things like vcp_printf(...) to debug stuff...  There are a few of these scattered in the code that are commented out now.

Update lib_f103.a